requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Yandex Disk integration
yadisk>=2.0.0

# Google Sheets integration
gspread>=5.0.0
google-auth>=2.0.0

# Document processing
python-docx>=1.0.0
pandas>=2.0.0
openpyxl>=3.0.0

# Image processing
Pillow>=10.0.0

# PDF processing (optional, if needed)
PyPDF2>=3.0.0
pdf2image>=1.16.0

# Utilities
requests>=2.31.0
